+++ frontend/src/components/tabs/MetricsTab.test.tsx
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import { Artifact, ArtifactType, Execution, Value } from '../src/mlmd/library';
+import { Artifact, ArtifactType, Execution, Value } from 'src/third_party/mlmd';
 import { render, waitFor } from '@testing-library/react';
 import { Struct } from 'google-protobuf/google/protobuf/struct_pb';
 import React from 'react';
-import * as mlmdUtils from 'src/lib/MlmdUtils';
+import * as mlmdUtils from 'src/mlmd/MlmdUtils';
 import { testBestPractices } from 'src/TestUtils';
 import { CommonTestWrapper } from 'src/TestWrapper';
 import { MetricsTab } from './MetricsTab';
