+++ frontend/src/components/tabs/MetricsTab.tsx
@@ -18,7 +18,7 @@ import * as React from 'react';
 import { useQuery } from 'react-query';
 import { ErrorBoundary } from 'src/atoms/ErrorBoundary';
 import { commonCss, padding } from 'src/Css';
-import { getArtifactTypes, getOutputArtifactsInExecution } from 'src/lib/MlmdUtils';
+import { getArtifactTypes, getOutputArtifactsInExecution } from 'src/mlmd/MlmdUtils';
 import { Artifact, ArtifactType, Execution } from 'src/third_party/mlmd';
 import Banner from '../Banner';
 import { MetricsVisualizations } from '../viewers/MetricsVisualizations';
